refactor(schemas): migrate Phase 21 endpoint schemas — 8 files, 24 classes (#6042)
- marketplace.py: migrate MarketplaceEntry, MarketplaceCatalogResponse,
  InstallRequest → schemas_workflows.py
- llm_awareness.py: migrate ContextRequest (dead code), PromptInjectionRequest,
  QueryAnalysisRequest → schemas_agent.py
- knowledge_graph_routes.py: migrate PipelineRunRequest, PipelineRunResponse,
  EventSearchRequest (dead code) → schemas_knowledge.py
- manual_mcp.py: migrate ManPageRequest, ManPageSearchRequest,
  ManPageResult (dead code) → schemas_code.py
- knowledge_search_aggregator.py: migrate UnifiedSearchRequest, ContextRequest,
  GraphRequest → schemas_knowledge.py
- knowledge_organization.py: migrate OrganizationKnowledgePolicy,
  OrganizationKnowledgeStats, UpdateOrganizationPolicyRequest → schemas_knowledge.py
- llm_optimization.py: migrate OptimizationRequest→LLMOptimizationRequest,
  ModelPerformanceData, InferenceOptimizationSettings → schemas_agent.py
- integration_github.py: migrate GitHubConnectionTestRequest, GitHubReviewRequest,
  GitHubCommentRequest → schemas_workflows.py

Renames to avoid generic-name collisions:
- OptimizationRequest → LLMOptimizationRequest

Bug fix (out-of-scope but blocking):
- knowledge_graph_routes.py used `parse_utc_iso` without importing it; added
  the import inside the search_events function (where it's used)

Cleanup: removed dead-code class imports (ContextRequest in llm_awareness,
EventSearchRequest in knowledge_graph_routes, ManPageResult in manual_mcp)
and unused DataResponse imports.
