+++ autobot-backend/api/schemas_code.py
@@ -2292,3 +2292,36 @@ class MCPRegistryStats(BaseModel):
     total_tools: int
     healthy_bridges: int
     last_updated: str
+
+
+# ---------------------------------------------------------------------------
+# manual_mcp.py schemas
+# ---------------------------------------------------------------------------
+
+
+class ManPageRequest(BaseModel):
+    """Request model for man page lookup."""
+
+    command: str = Field(..., description="Command name to look up (e.g. 'ls')")
+    section: Optional[str] = Field(None, description="Manual section (1-8). Defaults to section 1.")
+
+
+class ManPageSearchRequest(BaseModel):
+    """Request model for documentation index query."""
+
+    query: str = Field(..., description="Search query against the doc index")
+    max_results: int = Field(10, ge=1, le=50, description="Maximum results to return")
+
+
+class ManPageResult(BaseModel):
+    """Structured man page result."""
+
+    command: str
+    section: str
+    title: str
+    synopsis: str
+    description: str
+    options: str
+    examples: str
+    see_also: str
+    cached: bool
